bonus try

+++ src/display/rgb_calculations.c
@@ -31,10 +31,10 @@ t_rgb	multi_rgb(t_rgb rgb1, t_rgb rgb2, double ratio)
 
 t_rgb	add_rgb(t_rgb rgb1, t_rgb rgb2)
 {
-	t_rgb			r_rgb;
-	unsigned int    b;
-    unsigned int    g;
-    unsigned int    r;
+	t_rgb	r_rgb;
+	size_t	b;
+    size_t	g;
+    size_t	r;
 
 	r = rgb1.r + rgb2.r;
 	g = rgb1.g + rgb2.g;
